Fixed bug with "Save login info" context menu item, where a non-URL form action (such as "javascript:") leads to a JavaScript error. This was because the code assumed the action is a valid URL when looking at the results of a regular expression evaluation.
This is fixed by setting a value of "javascript:", as a special case for
actions with that scheme (as the Password Manager itself seems to do),
and checking that the regexp returns non-null in other cases; otherwise,
the form is rejected. I implemented this for the "Guess from current
page" button as well.

Fixes #16.